- correction of a lot a compilation warnings - correct posix type for read/write in order to improve portability - initial configuration manager instancier [ticket:323]

+++ src/common/dcap/gfal_dcap_plugin_bindings.c
@@ -83,9 +83,9 @@ gfal_file_handle gfal_dcap_openG(plugin_handle handle , const char* path, int fl
  * map to the libdcap read call
  * 
  */
-int gfal_dcap_readG(plugin_handle handle , gfal_file_handle fd, void* buff, size_t s_buff, GError** err){
+ssize_t gfal_dcap_readG(plugin_handle handle , gfal_file_handle fd, void* buff, size_t s_buff, GError** err){
 	gfal_plugin_dcap_handle h = (gfal_plugin_dcap_handle) handle;
-	int ret = h->ops->read(GPOINTER_TO_INT(fd->fdesc), buff, s_buff);
+	ssize_t ret = h->ops->read(GPOINTER_TO_INT(fd->fdesc), buff, s_buff);
 	if(ret <0)
 		dcap_report_error(h, __func__, err);
 	else
@@ -119,7 +119,7 @@ ssize_t gfal_dcap_pwriteG(plugin_handle handle , gfal_file_handle fd, const void
 	return ret;
 }
 
-int gfal_dcap_lseekG(plugin_handle handle , gfal_file_handle fd, off_t offset, int whence, GError** err){
+off_t gfal_dcap_lseekG(plugin_handle handle , gfal_file_handle fd, off_t offset, int whence, GError** err){
 	gfal_plugin_dcap_handle h = (gfal_plugin_dcap_handle) handle;
 	off_t ret = h->ops->lseek(GPOINTER_TO_INT(fd->fdesc), offset, (int) whence);
 	if(ret == ((off_t)0)-1)
@@ -129,9 +129,9 @@ int gfal_dcap_lseekG(plugin_handle handle , gfal_file_handle fd, off_t offset, i
 	return ret;
 }
 
-int gfal_dcap_writeG(plugin_handle handle , gfal_file_handle fd, const void* buff, size_t s_buff, GError** err){
+ssize_t gfal_dcap_writeG(plugin_handle handle , gfal_file_handle fd, const void* buff, size_t s_buff, GError** err){
 	gfal_plugin_dcap_handle h = (gfal_plugin_dcap_handle) handle;
-	int ret = h->ops->write(GPOINTER_TO_INT(fd->fdesc), buff, s_buff);
+	ssize_t ret = h->ops->write(GPOINTER_TO_INT(fd->fdesc), buff, s_buff);
 	if(ret <0)
 		dcap_report_error(h, __func__, err);
 	else
